requirements.txt
streamlit
pandas
plotly
openpyxl
python-dotenv
requests
openai